requirements.txt
colorama==0.4.4
openai==1.13.3
pandas==1.3.5
protobuf==3.20.3
Requests==2.31.0
tiktoken==0.5.1
yfinance==0.2.32
openpyxl==3.1.2
schedule==1.2.0
python-telegram-bot==20.7
python-dotenv==1.0.0
